requirements.txt
## This list is also used by setup.py - install_requires
## But there it removes all version restrictions
docopt
pyyaml
graphviz
